+++ tests/DBFactTest.php
@@ -180,4 +180,19 @@ class DBFactTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('TijsVerkoyen\DBFact\Types\Message', $response);
         $this->assertEquals('Ok', $response->KeepAlive);
     }
+
+	/**
+	 * Tests DBFact->receiveFile()
+	 */
+	public function testReceiveFile()
+	{
+		$files = array(
+			array(
+				'File' => file_get_contents('./order.xml'),
+				'FileName' => 'order.xml',
+			),
+		);
+		$response = $this->dbFact->receiveFile($files);
+		$this->assertTrue($response);
+	}
 }